Work around `DATABRICKS_HOST` normalisation issue during `install` and `uninstall` (#2052)
## Changes

This PR introduces a workaround for a `DATABRICKS_HOST` normalisation
issue in the underlying Python SDK (databricks/databricks-sdk-py#996).
We fixed this for the CLI entry points in #1706 but missed the `install`
and `uninstall` hooks, which this PR addresses.

### Relevant implementation details

The actual workaround is implemented in Blueprint, but a bit hard to get
to… the easiest way to get to it without modifying Blueprint is to
introduce the `App` subclass.

### Linked issues

Resolves #2037

### Functionality

- modified existing commands:

  - `databricks labs install lakebridge`
  - `databricks labs uninstall lakebridge`

### Tests

Manually tested, see logs below:

 - Normal situation, with `host=` set to just a host:

- [`databricks labs install lakebridge
--debug`](https://github.com/user-attachments/files/22517659/install-normal.log)
- [`databricks labs uninstall lakebridge
--debug`](https://github.com/user-attachments/files/22517616/uninstall-normal.log)

 - Situation with `host=` set to a login URL:

- [`databricks labs install lakebridge
--debug`](https://github.com/user-attachments/files/22517675/install-with-workaround.log)
- [`databricks labs uninstall lakebridge
--debug`](https://github.com/user-attachments/files/22517697/uninstall-with-workaround.log)