Fix type hints on LSP configuration classes and improve validation when loading (#2041)
## Changes

This PR updates type hints on the configuration classes, and implements
proper checking of the configuration during loading. Changes include:

- We now distinguish between missing attributes and attributes that are
present but malformed.
- With the options for a dialect, the prompt is now optional instead of
stubbing with an empty string. (The prompt is not needed for the `FIXED`
option type.)
- For options, the combinations of choices/default are now dealt with
properly depending on the kind of option it is.
 - Some tests that were using invalid fixtures have been fixed.

### Caveats/things to watch out for when reviewing

During deserialisation, in many places we now handle unhappy paths via
exceptions rather than checking in advance: the latter is unambiguous
whereas its easy to have subtle issues when the check doesn't quite
correctly anticipate or correspond to the actual usage.

### Relevant implementation details

Sadly, most of this has been done by hand even though there are
libraries that do this sort of thing for us.

### Tests

- existing unit tests