Switch from `urllib` to `requests` for making HTTP calls to PyPI and Maven Central (#2065)
## Changes

This PR updates our Maven and wheel installers to use the `requests`
package for making HTTP calls instead of Python's builtin `urllib`.
Although we try to avoid additional dependencies, it seems that Maven
Central's CDN (CloudFlare) is blocking HTTP calls include
`Python-urllib/3.xx` in the user-agent. We don't believe this is
specifically targeted at us: we are acting as a consumer application
(similar to maven/pip) and shouldn't be producing unreasonably amounts
of traffic.

### Caveats/things to watch out for when reviewing:

We still use the standard UA, but should extend it to identify
Databricks (and this application). I've added `TODO` markers for this as
an improvement.

### Linked issues

Resolves #2064 

### Functionality

- affects existing commands:

  - `databricks labs lakebridge install-transpile`
- `databricks labs install lakebridge` (if installed transpilers are
detected)

### Tests

- [X] manually tested
- [X] existing integration tests