Freshen project development dependencies (#2068)
## Changes

This PR:

 - Updates the project metadata.
- Freshens the development tooling that we use, along with minor
corresponding adjustments to the code to satisfy newer versions of the
tools.

### Relevant implementation details

Updated development tools:

 - Black (formatting)
 - Hatch
 - Mypy (static type checker)
 - Pytest (and the plugins we use: asyncio/cov/xdist)
 - Ruff (linter)

Pylint has been left alone, because it tends to be a bit more
complicated.

### Caveats/things to watch out for when reviewing

The updated version of mypy flagged a few minor issues in the code, but
these are reviewable by inspection.

### Tests

- manually tested
- existing unit tests
- existing integration tests