Update `transpile` to support `--overrides-path` and `--target-technology` arguments (#2072)
## Changes

This PR updates our `transpile` command to support some additional
arguments:

- `--overrides-file`: for specifying an overrides (JSON) file with
BB-based conversions.
- `--target-technology`: for specifying the target technology when using
BB to perform ETL conversions.

### Relevant implementation details

Adding support for these options uncovered some related problems
relating to the way dialect-specific options are configured and handled
at runtime. This meant some yak-shaving, including:

- During `transpile`, always checking the dialect-specific options even
if the configuration file is missing or invalid.
- Properly handling the `<none>` sentinel value used by BB at both
configuration time and `transpile`-time.
- Tidying up some of the unit test fixtures and tests that cover this
area of the code.

### Caveats/things to watch out for when reviewing:

Previously it looks like during `install-transpile` if the user hit
enter on the overrides-file option they would end up with the string
`"None"` as the value in the `transpiler_options` mapping. It's unclear
to me where this was ultimately dealt with, or whether these
configurations will just cause `transpile` to fail.

The integration tests here are flaky, and it's unclear why.
_[Late addition: I've just added some additional plumbing to the
integration tests to help with understanding why they're flaky. For some
reason the underlying `dbxconv` process sometimes terminates with 255 as
the exit code. Hopefully this will help understand why.]_

### Linked issues

Resolves #2070.

### Functionality

- [X] updated relevant user documentation
- [X] modified existing command: `databricks labs lakebridge install`

### Tests

- [X] manually tested
- [X] added unit tests
- [X] added integration tests