Ensure we output transpiled code (#2014)
### What does this PR do?
Currently, if there's a parsing error (usually, these mean the output is
unusable) we copy the input code into the output file. This leads to a
confusing interaction, because there is no indication that the file is
the same as the input, and the user is left wondering what happened.

This PR changes the output, so that the transpiled code, including all
issues we found, are available in the "transpiled" file.

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [x] manually tested
- [x] added unit tests
- [ ] added integration tests

---------

Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>
Co-authored-by: sundarshankar89 <sundar.shankar@databricks.com>